Refactor default options
